refactor: move session lifecycle functions to dedicated module
